% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignMatrix.R
\name{addSignMatrix}
\alias{addSignMatrix}
\title{Add a sign matrix to a graph object}
\usage{
addSignMatrix(graph)
}
\arguments{
\item{graph}{A \code{\link{graph}} object}
}
\value{
\code{\link{graph}} object with added sign matrix
}
\description{
Adding the sign matrix to the graph object. A sign matrix is
a matrix that indicate the action between nodes (+1 or -1, corresponding to activation and inhibition) 
in a network. A 0 means the two corresponding nodes are independent (not connected or activations/inhibitions 
cancel each other).
}
\seealso{
\code{\link{graph}}
}
\author{
Lars Snipen <lars.snipen@nmbu.no>
}
