% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr}
\alias{fdr}
\title{False discovery rate computation for GSEA.}
\usage{
fdr(NES, NESobs, NESnull)
}
\arguments{
\item{NES}{Normalised enrichment score for the given gene set.}

\item{NESobs}{Vector of normalised enrichment scores for all gene sets to be tested.}

\item{NESnull}{Matrix of normalised null distribution for all gene sets to be tested. Each column represent a gene set, and each row represent a permutation.}
}
\value{
FDR q-value for the given gene set.
}
\description{
Computes a false discovery rate (FDR) q-value for a gene
set given the observed normalised enrichment scores and a normalised null
distribution. Calculations are done separately for positive and negative
normalised enrichment scores.
}
\references{
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L.,
Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S.
and Mesirov, J. P (2005) Gene set enrichment analysis: A knowledge-based
approach for interpreting genome-wide expression pro?les, \emph{PNAS}, \bold{102},
15545-15550.
}
\author{
Guro Dorum
}
