% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDiffusion.R
\name{findDiffusion}
\alias{findDiffusion}
\title{Find diffusion matrix}
\usage{
findDiffusion(graph, alpha = 1)
}
\arguments{
\item{graph}{Graph object}

\item{alpha}{Parameter that controls speed of diffusion over a network}
}
\description{
Find the diffusion matrix based on a graph.
}
\author{
Solve Sabo <solve.sabo@nmbu.no>
}
