% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.rotation.R
\name{gsea.rotation}
\alias{gsea.rotation}
\title{GSEA rotation for longitudinal data}
\usage{
gsea.rotation(S, y, X, contrast, covmat = NULL, nrot = 10000,
  ES.p = 1)
}
\arguments{
\item{S}{Matrix indicating gene set membership with genes as rows and gene sets as columns. 1 indicates that the gene is a member, 0 indicates not a member.}

\item{y}{Matrix of gene expression data. Genes are columns and samples are rows}

\item{X}{Design matrix. Each row represent a sample}

\item{contrast}{Matrix of contrasts. Each column represent one contrast}

\item{covmat}{Covariance matrix (optional)}

\item{nrot}{Number of rotations}

\item{ES.p}{Weighting parameter for enrichment score (default=1)}
}
\value{
\item{Ngenes }{Size of gene sets}
\item{ES }{Enrichment score}
\item{NES }{Normalised enrichment score}
\item{p.value }{A p-value per gene set}
\item{q.value }{An FDR q-value for each gene set}
}
\description{
GSEA with rotation test for longitudinal data. Assumes common covariance matrix for all genes.
}
\examples{
#2 gene sets of 3 genes each. 3 time points. 1 replicate at each time point. 
#Total number of samples per gene is n
ngs <- 5
ng <- 3
nt <- 3
nrep <- 2
n <- nt*nrep

#Indicating which gene set each gene belongs to
iset <- rep(1:ngs, each=ng)
S <- matrix(0,ng*ngs,ngs)
for(i in 1:ngs) {
  S[iset==i,i] <- 1
}

#Design matrix per gene
Timer <- rep(1:nt,each=nrep) 
Time <- ordered(Timer)
#Design for random factors
Z <- cbind(Timer)
colnames(Z) <- c("Time")
#Design for fixed factors
X <- model.matrix(~Time)
X <- X[,-ncol(X)]
colnames(X) <- c("Gene","Linear")

#Set values for variance components
sigmat <- 2 #Time variance
sigmae <- 3 #Random error variance
phi <- 0.9 #Time correlation parameter
#Correlation between time points for same gene
Ts <- matrix(0,n,n)
for(i in 1:(n-1)) {
  for(j in (i+1):n) {
    Ts[i,j] <- abs(Z[j,1] - Z[i,1]) 
  }
}
Ts <- Ts + t(Ts)
R <- exp(-phi*Ts)
#Covariance matrix
subject <- rep(1,each=n)
V <- makecov(theta=c(sigmat,sigmae),correlation=R,design=Z,subject=subject)
Vroot <- chol(V)

#Simulate data with covariance structure
Y1 <- matrix(rnorm(n*ng*ngs),ng*ngs,n)
Y <- t(crossprod(t(Y1),Vroot)) 
#Add fixed gene effects and linear time effecs in first gene set
betaG <-  2; betaLT <- 1
Yeff <- Y
Yeff[,1:ng] <- Yeff[,1:ng] + matrix(X[,1]*betaG,ncol=ng,nrow=n) 
+ matrix(X[,2]*betaLT,ncol=ng,nrow=n)

#Estimate covariance matrix
#Constraints for optimisation and start values
A <- rbind(c(1,0), c(0,1))
b <- c(1e-6, 1e-6)
alphaStart <- c(sigmae/sigmat,phi)
reml_res <- reml( alphaStart=alphaStart, X=X, Y=Yeff, Z=Z, subject=subject, A=A, b=b)
sigmat.hat <- reml_res$sigma
sigmae.hat <- reml_res$alpha[1]*sigmat.hat
phi.hat <- reml_res$alpha[2]
R.hat <- exp(-phi.hat*Ts)
#Estimated covariance matrix
V.hat <- makecov(theta=c(sigmat.hat,sigmae.hat),correlation=R.hat,design=Z,subject=subject)

#Contrasts for testing gene effect and linear time effect
contrast <- matrix(0,ncol(X),2)
colnames(contrast) <- c("Gene","Linear")
contrast[colnames(X)=="Gene",1] <- 1
contrast[colnames(X)=="Linear",2] <- 1

#Contrast for testing gene effect
contrast <- matrix(0,ncol(X),2)
colnames(contrast) <- c("Gene","Linear")
contrast[colnames(X)=="Gene",1] <- 1
contrast[colnames(X)=="Linear",2] <- 1

#GSEA rotation
gsea.rotation(S=S, y=t(Yeff), X=X, contrast=contrast, covmat=V.hat, nrot=1000, ES.p=1)
}
\references{
Dorum, G., Snipen, L., Solheim, M. and Sabo (2009) Rotation Testing in Gene Set Enrichment
Analysis for Small Direct Comparison Experiments. \emph{Statistical Applications in Genetics
 and Molecular Biology}, \bold{8}(1), article 34.
 
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L.,
Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S.
and Mesirov, J. P (2005) Gene set enrichment analysis: A knowledge-based
approach for interpreting genome-wide expression profiles, \emph{PNAS}, \bold{102},
15545-15550.
}
\seealso{
\code{\link{gsea.rotation2}} for GSEA rotation that assumes common covariance matrix for all genes in the same gene set.
}
\author{
Guro Dorum
}
