\name{GSEArotTime-package}
\alias{GSEArotTime-package}
\alias{GSEArotTime}
\docType{package}
\title{
GSEA rotation for longitudinal expression data.
}
\description{
GSEA rotation for longitudinal expression data. 
}
\details{
\tabular{ll}{
Package: \tab GSEArotTime\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-26\cr
Depends: igraph
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Guro Dorum <guro.dorum@umb.no>
}
\references{
See Dorum et al. (2013)
}
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
