% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEArotTime-package.R
\docType{package}
\name{GSEArotTime}
\alias{GSEArotTime}
\alias{GSEArotTime-package}
\title{GSEArotTime: GSEA with rotation test for longitudinal data.}
\description{
A collection of functions for performing GSEA with rotation test on longitudinal data.
}
