% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecov.R
\name{makecov}
\alias{makecov}
\title{Make covariance matrix}
\usage{
makecov(theta, correlation, design, subject)
}
\arguments{
\item{theta}{A vector of variances for all random factors, where first element must be time series variance and second 
element must be the random error variance. The remaining elements can be variances for any other random design factors}

\item{correlation}{A matrix of time correlations between samples (usually zero for samples from different subjects)}

\item{design}{A nxp design matrix for the p random design factors. Must contain time effect in the first column. 
The time column must be numeric and the time intervals must be of actual size}

\item{subject}{A numeric vector indicating which subject each sample belongs to}
}
\value{
Covariance matrix
}
\description{
Construct a covariance matrix for a model with time series effect and other random effetcs. For balanced and unbalanced designs.
}
\seealso{
\code{\link{makecov2}} for construction of a covariance matrix that also includes gene dependencies.
}
\author{
Guro Dorum
}
