% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecov2.R
\name{makecov2}
\alias{makecov2}
\title{Make covariance matrix}
\usage{
makecov2(theta, Rt, Rg, design)
}
\arguments{
\item{theta}{A vector of variances for all random factors, where first element must be time series variance, second element must be gene variance, and third element must 
be the random error variance. The remaining elements can be variances for any other random design factors}

\item{Rt}{A matrix of time correlations between samples (usually zero for samples from different subjects)}

\item{Rg}{A matrix of correlations between genes}

\item{design}{A nxp design matrix for the p random design factors. Must contain time in the first column and gene in the second. 
The time column must be numeric and the time intervals must be of actual size.}
}
\value{
Covariance matrix
}
\description{
Construct a covariance matrix for model with time series correlation, gene correlation and correlations due to other random effetcs. For balanced and unbalanced designs.
}
\seealso{
\code{\link{makecov}} for construction of a covariance matrix without gene dependencies.
}
\author{
Guro Dorum
}
