% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.R
\name{QR}
\alias{QR}
\title{QR decomposition}
\usage{
QR(X)
}
\arguments{
\item{X}{Square matrix (nxn)}
}
\value{
\item{Q}{Orientation matrix}
\item{R}{Configuration matrix}
}
\description{
Function that performs a QR decomposition on a square matrix.
}
\author{
Solve Sabo
}
