% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reml.R
\name{reml}
\alias{reml}
\title{Restricted maximum likelihood to estimate the covariance matrix of longitudinal expression data with correlation between samples}
\usage{
reml(alphaStart, X, Y, Z, subject, A, b)
}
\arguments{
\item{alphaStart}{A numeric vector of start values for the parameters to be estimated. The first parameter must be random error variance, the last must be a 
parameter controlling correlation between time points. The remaining parameters can be variances for other random design factors}

\item{X}{Design matrix with fixed design variables}

\item{Y}{Matrix of observed values. Rows are samples and columns are variables.}

\item{Z}{Design matrix for random design factors}

\item{subject}{A numeric vector indicating which subject each sample belongs to}

\item{A}{Restrictions for parameters. See \code{\link{constrOptim}}}

\item{b}{Restrictions for parameters. See \code{\link{constrOptim}}}
}
\value{
\item{alpha.hat }{Vector of estimates for the parameters in alphaStart}
\item{beta.hat }{Vector of estimated parameters for fixed design factors}
\item{sigma }{Estimated common variance}
\item{maxlik }{The maximum likelihood value}
}
\description{
Restricted maximum likelihood for longitudinal expression data where genes are assumed independent. All genes are assumed to have identical covariance matrix.
}
\details{
Arguments \code{A} and \code{b} corresponds to arguments \code{ui} and \code{ci} in \code{\link{constrOptim}}.
}
\seealso{
\code{\link{reml2}} and \code{\link{reml_genedep}} for restricted maximum likelihood where gene correlations are assumed within gene set.
}
