% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reml2.R
\name{reml2}
\alias{reml2}
\title{Restricted maximum likelihood to estimate the covariance matrix of longitudinal expression data with gene dependencies within gene sets and correlation between samples}
\usage{
reml2(alphaStart, X, Y, Z, Ts, Gs, Bs = NULL, set, A, b)
}
\arguments{
\item{alphaStart}{A numeric vector of start values for the parameters to be estimated. The first parameter must be gene variance, the second parameter random error variance, 
the third must be a parameter controlling correlation between time points and the fourth must be a parameter controlling correlation between genes. The remaining parameters can 
be variances for other random design factors.}

\item{X}{Design matrix with fixed design variables for a gene set (assuming same design for all gene sets)}

\item{Y}{Vector of observed values for all gene sets}

\item{Z}{Design matrix for random design factors for a gene set (assuming same design for all gene sets)}

\item{Ts}{Structure of time dependencies between samples in a gene set (assuming same design for all gene sets)}

\item{Gs}{Structure of gene dependencies between samples in a gene set (assuming same design for all gene sets)}

\item{Bs}{(Optional) structure of batch dependencies between samples in a gene set (assuming same design for all gene sets)}

\item{set}{Vector of indices indicating gene set for each sample}

\item{A}{Restrictions for parameters. See \code{\link{constrOptim}}}

\item{b}{Restrictions for parameters. See \code{\link{constrOptim}}}
}
\value{
\item{alpha.hat }{Vector of estimates for the parameters in alphaStart}
\item{beta.hat }{Vector of estimated parameters for fixed design factors}
\item{sigma }{Estimated common variance}
\item{maxlik }{The maximum likelihood value}
}
\description{
Restricted maximum likelihood for longitudinal expression data where genes are assumed correlated within gene set and uncorrelated between different gene sets. 
All genes within the same gene set are assumed to have identical covariance matrix. Also requires that all gene sets are of same size and hence have identical covariance matrix.
}
\details{
Arguments \code{A} and \code{b} corresponds to arguments \code{ui} and \code{ci} in \code{\link{constrOptim}}.
}
\seealso{
\code{\link{reml}} for restricted maximum likelihood without gene correlations and \code{\link{reml_genedep}} that does not assume identical gene set size.
}
