% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation}
\alias{rotation}
\title{Rotations}
\usage{
rotation(Y, method)
}
\arguments{
\item{Y}{Matrix with expression values, where rows are samples and columns are genes.}

\item{method}{The method to be used for rotation, either 1 (all rotated genes have expectation 0) or 2 (non-zero means allowed)}
}
\value{
Matrix with rotated expression values
}
\description{
Rotates a gene expression matrix as a step for generating a null distribution for a given gene set.
}
\author{
Guro Dorum
}
