% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{significance}
\alias{significance}
\title{Significance level computation in GSEA}
\usage{
significance(ES, ESnull, doplot = FALSE)
}
\arguments{
\item{ES}{Observed enrichment score.}

\item{ESnull}{Null distribution.}

\item{doplot}{If \code{TRUE}, a histogram over the normalised null distribution is drawn. Default is \code{FALSE}.}
}
\value{
\item{p.value}{p-value for the gene set}
  \item{NES}{Normalised observed enrichment score}
 \item{NESnull}{Normalised estimated null distribution}
}
\description{
Computes a p-value for a gene set.
}
\details{
Function that computes a p-value for a gene set given an enrichment score 
and a null distribution. Calculations are done separately for positive and
negative enrichment scores. The function also returns normalised observed 
enrichment score and normalised null distribution for the gene set. 
In addition, a histogram over the normalised null distribution can be plotted.
}
\author{
Guro Dorum
}
